refactor(exec): extract review runner
